(refactor/deps): Add react-router-dom and refactor dir structure
