+++ src/components/App/App.css
@@ -3,7 +3,7 @@
 }
 
 .abc {
-  color: yellow;
+  color: blue;
 }
 
 .abcd {
